Time: 2 ms (36.59%) | Memory: 39.5 MB (48.78%) - LeetSync

+++ LeetCode/1461-count-all-valid-pickup-and-delivery-options/count-all-valid-pickup-and-delivery-options.java
@@ -1,11 +1,11 @@
 class Solution {
     public int countOrders(int n) {
-        double [] dp = new double [n + 1];
+        long [] dp = new long [n + 1];
         dp[1] = 1;
 
         for (int i = 2; i <= n; i++) {
-            double res = (dp[i - 1] * (sum(2 * (i - 1) + 1)));
-            dp[i] = res % (Math.pow(10, 9) + 7);
+            long res = (dp[i - 1] * (sum(2 * (i - 1) + 1)));
+            dp[i] = (long) (res % (Math.pow(10, 9) + 7));
         } //for
 
         return (int) (dp[n] % (Math.pow(10, 9) + 7));
